Fix windows compile warnings (#451)
Fix windows compile warnings and from now all warnings are treated as errors.

+++ src/comm/tcp_server.cpp
@@ -121,7 +121,7 @@ void TCPServer::shutdown()
   memset(&address, 0, sizeof(address));
   address.sin_family = AF_INET;
   address.sin_addr.s_addr = htonl(INADDR_LOOPBACK);
-  address.sin_port = htons(port_);
+  address.sin_port = htons(static_cast<uint16_t>(port_));
 
   ::connect(shutdown_socket, reinterpret_cast<const sockaddr*>(&address), sizeof(address));
 
@@ -152,7 +152,7 @@ void TCPServer::bind(const size_t max_num_tries, const std::chrono::milliseconds
 
   // INADDR_ANY is a special constant that signalizes "ANY IFACE",
   server_addr.sin_addr.s_addr = htonl(INADDR_ANY);
-  server_addr.sin_port = htons(port_);
+  server_addr.sin_port = htons(static_cast<uint16_t>(port_));
   int err = -1;
   size_t connection_counter = 0;
   do
